Relocate SignalEvent + map_event_to_signal to api-model (#377)
SignalEvent and the api::Event -> SignalEvent projection move from
client-common into api-model/src/signal.rs, so the wasm-targeting client cores
(client-ui-common, the web SPA) can consume the signal stream and convert wire
events without client-common's native transport tail. client-common's signal.rs
and ws_client::map_event_to_signal become re-exports, so existing
client_common::SignalEvent / ws_client::map_event_to_signal paths are unchanged.

api-model (incl. the new signal module) compiles to wasm32-unknown-unknown;
full native workspace builds; fmt + clippy --all-targets -D warnings clean;
protocol/api-model/client-common tests + core render_blurb tests pass.
