+++ crates/api-model/src/lib.rs
@@ -10,6 +10,9 @@ use std::collections::BTreeMap;
 
 use serde::{Deserialize, Serialize};
 
+pub mod signal;
+pub use signal::{SignalEvent, map_event_to_signal};
+
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(rename_all = "snake_case")]
 pub enum Command {
