+++ crates/client-common/src/ws_client.rs
@@ -388,127 +388,11 @@ fn build_tls_connector(ca_cert_path: Option<&Path>) -> Result<tokio_tungstenite:
     Ok(tokio_tungstenite::Connector::Rustls(Arc::new(config)))
 }
 
-pub fn map_event_to_signal(event: api::Event) -> Option<SignalEvent> {
-    match event {
-        api::Event::UserMessageAdded {
-            conversation_id,
-            request_id,
-            content,
-        } => Some(SignalEvent::UserMessageAdded {
-            conversation_id,
-            request_id,
-            content,
-        }),
-        api::Event::AssistantDelta {
-            conversation_id,
-            request_id,
-            chunk,
-        } => Some(SignalEvent::Chunk {
-            conversation_id,
-            request_id,
-            chunk,
-        }),
-        api::Event::AssistantCompleted {
-            conversation_id,
-            request_id,
-            full_response,
-        } => Some(SignalEvent::Complete {
-            conversation_id,
-            request_id,
-            full_response,
-        }),
-        api::Event::AssistantError {
-            conversation_id,
-            request_id,
-            error,
-        } => Some(SignalEvent::Error {
-            conversation_id,
-            request_id,
-            error,
-        }),
-        api::Event::ConversationTitleChanged {
-            conversation_id,
-            title,
-        } => Some(SignalEvent::TitleChanged {
-            conversation_id,
-            title,
-        }),
-        api::Event::ConversationListChanged { conversation_id } => {
-            Some(SignalEvent::ConversationListChanged { conversation_id })
-        }
-        api::Event::AssistantStatus {
-            conversation_id,
-            request_id,
-            message,
-        } => Some(SignalEvent::Status {
-            conversation_id,
-            request_id,
-            message,
-        }),
-        api::Event::ContextUsage {
-            conversation_id,
-            request_id,
-            used_tokens,
-            budget_tokens,
-            compaction_active,
-        } => Some(SignalEvent::ContextUsage {
-            conversation_id,
-            request_id,
-            used_tokens,
-            budget_tokens,
-            compaction_active,
-        }),
-        api::Event::ConfigChanged { .. } => None,
-        api::Event::ConversationWarningEmitted {
-            conversation_id,
-            warning,
-        } => Some(SignalEvent::ConversationWarning {
-            conversation_id,
-            warning,
-        }),
-        // Background-task events (issue #110) — surfaced verbatim on the
-        // signal channel so process-manager UIs (adele-tui#45, adele-gtk
-        // follow-up) can react. The TaskView/TaskLogEntry types are
-        // re-exported from `api-model`; clients consume them directly.
-        api::Event::TaskStarted { task } => Some(SignalEvent::TaskStarted { task }),
-        api::Event::TaskProgress { id, progress_hint } => {
-            Some(SignalEvent::TaskProgress { id, progress_hint })
-        }
-        api::Event::TaskLogAppended { id, entry } => {
-            Some(SignalEvent::TaskLogAppended { id, entry })
-        }
-        api::Event::TaskCompleted {
-            id,
-            status,
-            last_error,
-        } => Some(SignalEvent::TaskCompleted {
-            id,
-            status,
-            last_error,
-        }),
-        api::Event::ScratchpadChanged { conversation_id } => {
-            Some(SignalEvent::ScratchpadChanged { conversation_id })
-        }
-        // Client-side tool execution (#107/#231): surfaced on the signal
-        // stream so a client that advertised client-local tools (voice first)
-        // can execute the requested tool and post the result back via
-        // `Connector::submit_client_tool_result`. The `TaskId` newtype is
-        // unwrapped to its inner `String` to match the rest of this stream.
-        api::Event::ClientToolCall {
-            task_id,
-            conversation_id,
-            tool_call_id,
-            tool_name,
-            arguments,
-        } => Some(SignalEvent::ClientToolCall {
-            task_id: task_id.0,
-            conversation_id,
-            tool_call_id,
-            tool_name,
-            arguments,
-        }),
-    }
-}
+// `SignalEvent` and the `api::Event` → `SignalEvent` projection now live in
+// `api-model` (so the wasm client cores can use them without this crate's native
+// transport tail, #377). Re-exported here so `crate::ws_client::map_event_to_signal`
+// callers (uds_client, the streaming loop above) are unchanged.
+pub use desktop_assistant_api_model::signal::map_event_to_signal;
 
 #[cfg(test)]
 mod tests {
